Move FlashcardInstruction into 'flashcards' directory. Wrap checkbox in label.
